pyproject.toml
[tool.poetry]
name = "python-gRPC-asycnio"
version = "0.1.0"
description = ""
authors = ["ErEr <jeffliao92316046@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
grpcio-tools = "^1.67.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
